A few tweaks
